行业：etf,计算区间涨幅

+++ src/main/java/ttjj/dto/RankBizDataDiff.java
@@ -44,7 +44,7 @@ public class RankBizDataDiff {
     /**
      * f2 指数点（当前价格）
      */
-    private Double f2;
+    private BigDecimal f2;
     /**
      * f3 涨跌幅
      */
@@ -93,19 +93,19 @@ public class RankBizDataDiff {
     /**
      * f15 最高
      */
-    private Double f15;
+    private BigDecimal f15;
     /**
      * f16 最低
      */
-    private Double f16;
+    private BigDecimal f16;
     /**
      * f17 今日开盘
      */
-    private Double f17;
+    private BigDecimal f17;
     /**
      * f18 昨日收盘
      */
-    private Double f18;
+    private BigDecimal f18;
 
     /**
      * f19 ？2-板块；9-沪市；10-深市：
@@ -495,11 +495,11 @@ public class RankBizDataDiff {
         this.f1 = f1;
     }
 
-    public Double getF2() {
+    public BigDecimal getF2() {
         return f2;
     }
 
-    public void setF2(Double f2) {
+    public void setF2(BigDecimal f2) {
         this.f2 = f2;
     }
 
@@ -599,35 +599,35 @@ public class RankBizDataDiff {
         this.f14 = f14;
     }
 
-    public Double getF15() {
+    public BigDecimal getF15() {
         return f15;
     }
 
-    public void setF15(Double f15) {
+    public void setF15(BigDecimal f15) {
         this.f15 = f15;
     }
 
-    public Double getF16() {
+    public BigDecimal getF16() {
         return f16;
     }
 
-    public void setF16(Double f16) {
+    public void setF16(BigDecimal f16) {
         this.f16 = f16;
     }
 
-    public Double getF17() {
+    public BigDecimal getF17() {
         return f17;
     }
 
-    public void setF17(Double f17) {
+    public void setF17(BigDecimal f17) {
         this.f17 = f17;
     }
 
-    public Double getF18() {
+    public BigDecimal getF18() {
         return f18;
     }
 
-    public void setF18(Double f18) {
+    public void setF18(BigDecimal f18) {
         this.f18 = f18;
     }
